Move SOR data into data module.
