+++ balancer-js/examples/swapLocalFork.ts
@@ -6,8 +6,8 @@ import dotenv from 'dotenv';
 import { BalancerSDK, Network } from '../src/index';
 import mainnetTop10 from '@/test/lib/mainnet-top-10.json';
 import { MockPoolDataService } from '@/test/lib/mockPool';
-import { getOnChainBalances } from '@/modules/sor/pool-data/onChainData';
-import { mapPools } from '@/modules/sor/pool-data/subgraphPoolDataService';
+import { getOnChainBalances } from '@/modules/data/pool/onChainData';
+import { mapPools } from '@/modules/data/pool/subgraphPoolDataService';
 import { JsonRpcProvider } from '@ethersproject/providers';
 import { Wallet } from '@ethersproject/wallet';
 import { getNetworkConfig } from '../src/modules/sdk.helpers';
